fix: improve Windows and macOS compatibility and safety
- Define custom processor_power_information struct for Windows.
- Fix Windows storage enum values and macro usage.
- Remove global scope qualifiers in macOS and POSIX code.
- Add WiFi radio state buffer size check.
- Update tests and allow manual CI trigger.

+++ include/syscape/detail/wifi/common.hpp
@@ -209,7 +209,7 @@ frequency_to_channel(std::uint32_t freq_mhz) noexcept {
             return static_cast<std::uint16_t>((freq_mhz - 2407U) / 5U);
         }
     } else if (freq_mhz == 2484U) {
-        return 14U;
+        return static_cast<std::uint16_t>(14U);
     } else if (freq_mhz >= 5000U && freq_mhz <= 5895U) {
         if ((freq_mhz - 5000U) % 5U == 0U) {
             return static_cast<std::uint16_t>((freq_mhz - 5000U) / 5U);
